requirements.txt
pygame>=2.5
Pillow>=10.0